refactor: split scheduled learning replay tests
